Add test for AddTutorialCommand
